refactor: Move `is` to `traversing`, optimize
Moved so it can use `getFilterFn`.
